# shared/generated/swagger.json
{
	"components": {
...
		"schemas": {
			"Town": {
				"description": "The public-facing model that represents a town. More information about the town\nis available for users who join it.",
				"properties": {
					"friendlyName": {
						"type": "string",
						"description": "The name that users see on the landing page to determine which town to join"
					},
					"townID": {
						"type": "string",
						"description": "An internal ID that is used to uniquely identify each town"
					},
					"currentOccupancy": {
						"type": "number",
						"format": "double",
						"description": "The current number of players in this town"
					},
					"maximumOccupancy": {
						"type": "number",
						"format": "double",
						"description": "The maximum number of players allowed in this town"
					}
				},
				"required": [
					"friendlyName",
					"townID",
					"currentOccupancy",
					"maximumOccupancy"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"InvalidParametersError": {
				"properties": {
					"code": {},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"TownSettingsUpdate": {
				"properties": {
					"isPubliclyListed": {
						"type": "boolean"
					},
					"friendlyName": {
						"type": "string"
					}
				},
				"type": "object"
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "covey-town-townService",
		"version": "1.0.0",
		"description": "Backend service for processing Covey.Town town requests",
		"license": {
			"name": "BSD-3-Clause"
		},
		"contact": {
			"name": "Jonathan Bell"
		}
	},
	"openapi": "3.0.0",
	"paths": {
		"/towns": {
			"get": {
				"operationId": "ListTowns",
				"responses": {
					"200": {
						"description": "list of towns",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Town"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "List all towns that are set to be publicly available",
				"tags": [
					"towns"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreateTown",
				"responses": {
					"200": {
						"description": "The ID of the newly created town, and a secret password that will be needed to update or delete this town.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TownCreateResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"townID": "stringID",
											"townUpdatePassword": "secretPassword"
										}
									}
								}
							}
						}
					}
				},
				"description": "Create a new town",
				"tags": [
					"towns"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"description": "The public-facing information for the new town",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/TownCreateParams",
								"description": "The public-facing information for the new town"
							},
							"example": {
								"friendlyName": "My testing town public name",
								"isPubliclyListed": true
							}
						}
					}
				}
			}
		},
		"/towns/{townID}": {
			"patch": {
				"operationId": "UpdateTown",
				"responses": {
					"204": {
						"description": "No content"
					},
					"400": {
						"description": "Invalid password or update values specified",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/InvalidParametersError"
								}
							}
						}
					}
				},
				"description": "Updates an existing town's settings by ID",
				"tags": [
					"towns"
				],
				"security": [],
				"parameters": [
					{
						"description": "town to update",
						"in": "path",
						"name": "townID",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "town update password, must match the password returned by createTown",
						"in": "header",
						"name": "X-CoveyTown-Password",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"description": "The updated settings",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/TownSettingsUpdate",
								"description": "The updated settings"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteTown",
				"responses": {
					"204": {
						"description": "No content"
					},
					"400": {
						"description": "Invalid password or update values specified",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/InvalidParametersError"
								}
							}
						}
					}
				},
				"description": "Deletes a town",
				"tags": [
					"towns"
				],
				"security": [],
				"parameters": [
					{
						"description": "ID of the town to delete",
						"in": "path",
						"name": "townID",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "town update password, must match the password returned by createTown",
						"in": "header",
						"name": "X-CoveyTown-Password",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/towns/{townID}/conversationArea": {
			"post": {
				"operationId": "CreateConversationArea",
				"responses": {
					"204": {
						"description": "No content"
					},
					"400": {
						"description": "Invalid values specified",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/InvalidParametersError"
								}
							}
						}
					}
				},
				"description": "Creates a conversation area in a given town",
				"tags": [
					"towns"
				],
				"security": [],
				"parameters": [
					{
						"description": "ID of the town in which to create the new conversation area",
						"in": "path",
						"name": "townID",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "session token of the player making the request, must match the session token returned when the player joined the town",
						"in": "header",
						"name": "X-Session-Token",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"description": "The new conversation area to create",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ConversationArea",
								"description": "The new conversation area to create"
							}
						}
					}
				}
			}
		},
		"/towns/{townID}/viewingArea": {
			"post": {
				"operationId": "CreateViewingArea",
				"responses": {
					"204": {
						"description": "No content"
					},
					"400": {
						"description": "Invalid values specified",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/InvalidParametersError"
								}
							}
						}
					}
				},
				"description": "Creates a viewing area in a given town",
				"tags": [
					"towns"
				],
				"security": [],
				"parameters": [
					{
						"description": "ID of the town in which to create the new viewing area",
						"in": "path",
						"name": "townID",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "session token of the player making the request, must\nmatch the session token returned when the player joined the town",
						"in": "header",
						"name": "X-Session-Token",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"description": "The new viewing area to create",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ViewingArea",
								"description": "The new viewing area to create"
							}
						}
					}
				}
			}
		},
		"/towns/{townID}/karaokeArea": {
			"post": {
				"operationId": "CreateKaraokeArea",
				"responses": {
					"204": {
						"description": "No content"
					},
					"400": {
						"description": "Invalid values specified",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/InvalidParametersError"
								}
							}
						}
					}
				},
				"description": "Creates a karaoke area in a given town",
				"tags": [
					"towns"
				],
				"security": [],
				"parameters": [
					{
						"description": "ID of the town in which to create the new karaoke area",
						"in": "path",
						"name": "townID",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "session token of the player making the request, must\nmatch the session token returned when the player joined the town",
						"in": "header",
						"name": "X-Session-Token",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"description": "The new karaoke area to create",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/KaraokeArea",
								"description": "The new karaoke area to create"
							}
						}
					}
				}
			}
		},
		"/towns/{townID}/songRating": {
			"patch": {
				"operationId": "UpdateSongRating",
				"responses": {
					"204": {
						"description": "No content"
					},
					"400": {
						"description": "Invalid values specified",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/InvalidParametersError"
								}
							}
						}
					}
				},
				"description": "Updates the current song rating, or add a rating to an unrated song",
				"tags": [
					"towns"
				],
				"security": [],
				"parameters": [
					{
						"description": "ID of the town in which to update the karaoke area image contents",
						"in": "path",
						"name": "townID",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "ID of the song to update",
						"in": "query",
						"name": "songID",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "rating",
						"required": true,
						"schema": {
							"type": "number",
							"enum": [
								1,
								2,
								3,
								4,
								5
							]
						}
					},
					{
						"description": "session token of the player making the request, must\nmatch the session token returned when the player joined the town",
						"in": "header",
						"name": "X-Session-Token",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/towns/authorize": {
			"get": {
				"operationId": "SpotifyAuthorize",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"towns"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "codeChallenge",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "state",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/towns/callback": {
			"get": {
				"operationId": "SpotifyCallback",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"towns"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "codeVerifier",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	],
...
}